added print

+++ symtablehash.c
@@ -186,7 +186,7 @@ void SymTable_expand(SymTable_T oSymTable)
             printf("%d", hvalue);
             printf("\n");
             pnext = cur->next;
-            if (i != hvalue)
+            if ((int)i != hvalue)
             {
                 printf("reallocated ");
                 printf("%s", cur->key);
